Changes to sparsing functions and MFCC extraction

+++ models/AR_ASLG_HMM.py
@@ -1516,7 +1516,7 @@ class AR_ASLG_HMM:
             n_rows = 1
         else:
             n_cols = 6
-            n_rows = np.ceil(n_plots/6)
+            n_rows = int(np.ceil(n_plots/6))
         plt.figure("Scatter plots  ")
         plt.clf()
         for i in range(self.N):
